+++ crates/core/src/data_formats/mod.rs
@@ -1,4 +1,5 @@
 pub mod json;
 pub mod sql;
+mod util;
 pub mod xml;
 pub mod yaml;